+++ src/EventSourcing/tests/Logitar.EventSourcing.SqlServer.IntegrationTests/Init.sql
@@ -1,7 +1,20 @@
+IF OBJECT_ID(N'[__EFMigrationsHistory]') IS NOT NULL
+BEGIN
+    DROP TABLE [__EFMigrationsHistory];
+END;
+GO
+
 IF OBJECT_ID(N'[Events]') IS NOT NULL
 BEGIN
     DROP TABLE [Events];
 END;
+GO
+
+CREATE TABLE [__EFMigrationsHistory] (
+    [MigrationId] nvarchar(150) NOT NULL,
+    [ProductVersion] nvarchar(32) NOT NULL,
+    CONSTRAINT [PK___EFMigrationsHistory] PRIMARY KEY ([MigrationId])
+);
 GO
 
 BEGIN TRANSACTION;
@@ -11,9 +24,9 @@ CREATE TABLE [Events] (
     [EventId] bigint NOT NULL IDENTITY,
     [Id] uniqueidentifier NOT NULL,
     [ActorId] nvarchar(255) NOT NULL,
+    [IsDeleted] bit NULL,
     [OccurredOn] datetime2 NOT NULL,
     [Version] bigint NOT NULL,
-    [DeleteAction] nvarchar(255) NOT NULL,
     [AggregateType] nvarchar(255) NOT NULL,
     [AggregateId] nvarchar(255) NOT NULL,
     [EventType] nvarchar(255) NOT NULL,
@@ -37,10 +50,17 @@ GO
 CREATE UNIQUE INDEX [IX_Events_Id] ON [Events] ([Id]);
 GO
 
+CREATE INDEX [IX_Events_IsDeleted] ON [Events] ([IsDeleted]);
+GO
+
 CREATE INDEX [IX_Events_OccurredOn] ON [Events] ([OccurredOn]);
 GO
 
 CREATE INDEX [IX_Events_Version] ON [Events] ([Version]);
+GO
+
+INSERT INTO [__EFMigrationsHistory] ([MigrationId], [ProductVersion])
+VALUES (N'20230804163434_CreateEventTable', N'7.0.9');
 GO
 
 COMMIT;